+++ src/app/pages/entries/entry-form/entry-form.component.ts
@@ -4,7 +4,7 @@ import { Entry } from 'src/app/models/entry.model';
 import { Category } from 'src/app/models/category.model';
 import { CategoryService } from 'src/app/services/category.service';
 import { EntryService } from 'src/app/services/entry.service';
-import { BaseResourceFormComponent } from 'src/app/shared/components/base-resource-form.component';
+import { BaseResourceFormComponent } from 'src/app/shared/components/base-resource-form/base-resource-form.component';
 
 @Component({
   selector: 'app-entry-form',
